User entries + clear
Input fields now work I had to add a new input for the value, and also the clear button works, although it has to reload the page once it clears in order to get a new list done - simpler than deleting current table and creating a new one.

next steps would be to add a button to delete a specific localStorage item.